+++ packages/odata-service-inquirer/src/error-handler/help/images/index.ts
@@ -1,0 +1,1 @@
+export * from './guidedAnswersIcon_svg_base64';